gateway/requirements.txt
grpcio==1.66.1
grpcio-tools==1.66.1
PyYAML==6.0.2
